% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oqr_ps.R
\name{read_oqr_ps}
\alias{read_oqr_ps}
\title{Read OQR data}
\usage{
read_oqr_ps(folder, quarter, joinMOU = FALSE, ps = TRUE, impute = TRUE)
}
\arguments{
\item{folder}{data location prior to quarter specification}

\item{quarter}{data quarter, used in path and elsewhere}

\item{joinMOU}{whether the data should be joined the moulist file. needed for NA measures for non-respondents}

\item{ps}{whether the data includes population data; true by default}

\item{impute}{whether OP-3b observation should be imputed if OP-2 is present; true by default}
}
\description{
Used for reading in OQR data in the Annual Reports
}
